requirements.txt
flask
moviepy
openai-whisper
torch
numpy
Pillow
